misc bugs
